spec-demo-code: Support the new SPEC vendor/device ID
This enlarges the pci_id table to support three vendor/device pairs
instead of 1.  One pair is the gennum one, the other is the new
spec identifier and the third is the user-defined on (vendor=, device=).

This derails from the original idea of having a purely-generic driver,
but it allows to drive an older and a newer spec at the same time.